Moved recursive dependency finder to own module
